delete option for admin

+++ controllers/api/timesheet.controller.js
@@ -77,13 +77,27 @@ function getTimesheet(req, res){
 
 function deleteTimesheet(req, res){
     var userId = req.user.sub;
-    timesheetService.deleteTimesheet(req.params.id, userId)
-        .then(function (response) {
-            res.send(response);
-        })
-        .catch(function (err) {
-            res.status(400).send(err);
-        });
+    userService.getById(req.user.sub).then(function(currentUser) {
+        if (currentUser && currentUser.admin === true) {
+            timesheetService.adminDeleteTimesheet(req.params.id)
+                .then(function (response) {
+                    res.send(response);
+                })
+                .catch(function (err) {
+                    res.status(400).send(err);
+                });
+        } else {
+            timesheetService.deleteTimesheet(req.params.id, userId)
+                .then(function (response) {
+                    res.send(response);
+                })
+                .catch(function (err) {
+                    res.status(400).send(err);
+                });
+        }
+    }).catch(function(err) {
+        res.status(400).send(err);
+    });
 }
 
 function getMyReport(req, res) {
